Set AutoReqProv to No for now
The new `mingw.attr` Requires generator in clr-rpm-config works as expected,
but the Requires list now includes symbols provided by `binutils-lib`, which is
unexpected. Formerly, those Requires were not generated because
`mingw-binutils` Provided the same elfdeps symbols as `binutils-lib`,
eliminating the need for Requiring them.

The versioned libraries, like `libbfd-2.36.1.so`, will pose a problem for
updating this package after the next binutils upgrade, since we would
temporarily need a binutils compat package to bootstrap it again.

For now, set "AutoReqProv: No" to avoid the need to bootstrap.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>